HPL1: Fix Signed vs. Unsigned Comparison GCC Warning in Engine String Class

+++ engines/hpl1/engine/system/String.cpp
@@ -575,7 +575,7 @@ tStringVec &cString::GetStringVec(const tString &asData, tStringVec &avVec, tStr
 		} else {
 			start = true;
 			str += c;
-			if (i == asData.length() - 1)
+			if (i == (int)asData.length() - 1)
 				avVec.push_back(str);
 		}
 	}
